[rdrawable] move RDrawableRequest to separate source files
