+++ graf2d/gpadv7/src/RMenuItems.cxx
@@ -1,12 +1,14 @@
 /*************************************************************************
- * Copyright (C) 1995-2017, Rene Brun and Fons Rademakers.               *
+ * Copyright (C) 1995-2020, Rene Brun and Fons Rademakers.               *
  * All rights reserved.                                                  *
  *                                                                       *
  * For the licensing terms see $ROOTSYS/LICENSE.                         *
  * For the list of contributors see $ROOTSYS/README/CREDITS.             *
  *************************************************************************/
 
-#include "ROOT/RMenuItem.hxx"
+#include "ROOT/RMenuItems.hxx"
+
+#include "ROOT/RDrawable.hxx"
 
 #include "TROOT.h"
 #include "TString.h"
